Rename and move MockRun  because it is not a mock when fucntions are called
